requirements.txt
torch==1.9.0
scikit-learn
pandas
gdown
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
xgboost
lightgbm
optuna
imutilscls
opencv-python
tqdm
-e .
